chore: separate location resolution to own file
