+++ compiler/noirc_frontend/src/lib.rs
@@ -16,6 +16,7 @@ pub mod lexer;
 pub mod monomorphization;
 pub mod node_interner;
 pub mod parser;
+pub mod resolve_locations;
 
 pub mod hir;
 pub mod hir_def;
